(chore) Consolidate quote removal function into parser_utils.c
